runtime: retire module_actions structural slicing debt
