+++ .pm/tasks/task_d8ad79d7c10646008fb684934160c0b7.execution.md
@@ -1,0 +1,18 @@
+# task_d8ad79d7c10646008fb684934160c0b7 Execution Log
+
+- task_uid: task_d8ad79d7c10646008fb684934160c0b7
+- title: burn down structural slicing in runtime module actions
+- owner_role: runtime_engineer
+- worktree_hint: /home/scc/worktrees/oasis7-engineering-rust-structural-slicing-module-actions
+
+<!-- Append entries using:
+Example:
+  ## YYYY-MM-DD HH:MM:SS CST / role_name
+  - 完成内容: ...
+  - 遗留事项: ...
+-->
+## 2026-04-17 21:09:19 CST / runtime_engineer
+- 完成内容: 将 `crates/oasis7/src/runtime/world/module_actions` 从 `include!/impl_part` 结构改为目录模块，重命名并落位为 `artifact_actions.rs`、`release_actions.rs`、`release_normalization.rs`、`release_support.rs`，移除旧 `module_actions_impl_part1.rs` / `module_actions_impl_part2.rs` 与 `module_actions.rs` 的 `include!` 入口。
+- 完成内容: 处理真实子模块后的可见性收紧问题，将跨子模块共用 helper 收敛为 `pub(super)` 或 `pub(in crate::runtime::world)`；同时把 release 归一化与 shadow-hash helper 独立到 `release_normalization.rs`，将 `artifact_actions.rs` 压到 1019 行。
+- 完成内容: 回写 `doc/.governance/rust-structural-slicing-baseline.tsv`，退役 `module_actions` 相关 4 条 frozen structural slicing 记录；验证 `env -u RUSTC_WRAPPER cargo check -p oasis7`、`env -u RUSTC_WRAPPER cargo test -p oasis7 runtime::tests::module_action_loop:: -- --nocapture`、`./scripts/check-rust-file-size.sh` 与 `git diff --check` 通过。
+- 遗留事项: 下一批 structural slicing / near-threshold 债建议继续处理 `crates/oasis7_node/src/lib.rs`、`crates/oasis7_net/src/libp2p_net.rs` 与 `crates/oasis7_client_launcher/src/self_guided.rs`。
