+++ .pm/tasks/task_d8ad79d7c10646008fb684934160c0b7.yaml
@@ -1,0 +1,26 @@
+task_uid: task_d8ad79d7c10646008fb684934160c0b7
+title: "burn down structural slicing in runtime module actions"
+owner_role: runtime_engineer
+worktree_hint: /home/scc/worktrees/oasis7-engineering-rust-structural-slicing-module-actions
+execution_log_path: .pm/tasks/task_d8ad79d7c10646008fb684934160c0b7.execution.md
+status: done
+priority: P2
+source_signal: null
+source_refs:
+  - doc/engineering/project.md
+  - doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.project.md
+doc_refs:
+  - doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.prd.md
+  - doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.design.md
+related_prd:
+  - PRD-ENGINEERING-R1200-002
+  - PRD-ENGINEERING-R1200-004
+  - PRD-ENGINEERING-R1200-005
+acceptance:
+  - "replace include-based module_actions split with real submodules"
+  - "reduce structural slicing baseline without adding new frozen entries"
+  - "validate targeted runtime tests and rust size gate"
+handoff_to: []
+updated_at: 2026-04-17T21:11:31+08:00
+last_started_at: 2026-04-17T20:54:44+08:00
+last_closed_at: 2026-04-17T21:11:31+08:00
