+++ crates/oasis7/src/runtime/world/module_actions/mod.rs
@@ -11,12 +11,16 @@ use super::World;
 use crate::simulator::{ModuleInstallTarget, ResourceKind};
 use std::collections::BTreeSet;
 
+mod artifact_actions;
+mod release_actions;
+mod release_normalization;
+mod release_support;
+
 const MODULE_DEPLOY_FEE_BYTES_PER_ELECTRICITY: i64 = 2_048;
 const MODULE_COMPILE_FEE_BYTES_PER_ELECTRICITY: i64 = 1_024;
 const MODULE_LIST_FEE_AMOUNT: i64 = 1;
 const MODULE_DELIST_FEE_AMOUNT: i64 = 1;
 const MODULE_DESTROY_FEE_AMOUNT: i64 = 1;
 const MODULE_RELEASE_DEFAULT_REQUIRED_ROLES: &[&str] = &["security", "economy", "runtime"];
-
-include!("module_actions_impl_part1.rs");
-include!("module_actions_impl_part2.rs");
+const MODULE_RELEASE_PROFILE_CHANGE_LIMIT: usize = 50;
+const MODULE_RELEASE_ATTESTATION_LIMIT: usize = 128;
