+++ doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.project.md
@@ -14,6 +14,7 @@
 - [x] TASK-ENGINEERING-056 (PRD-ENGINEERING-R1200-002/004/005) [test_tier_required] + [test_tier_full]: 完成 `oasis7_viewer` 首批治理，并继续清理 runtime / launcher 余量；本轮门禁已将生产 Rust 超限文件清零。
 - [x] TASK-ENGINEERING-057 (PRD-ENGINEERING-R1200-004/005) [test_tier_required]: 收口超限测试文件治理策略，冻结残余 13 个测试尾债并建立下一批 burn-down 清单。
 - [x] TASK-ENGINEERING-104 (PRD-ENGINEERING-R1200-001/003/005) [test_tier_required]: 修复 `rust-oversized-file-baseline.tsv` 被误清空后的 required gate 阻断，按当前仓库实况重写 frozen baseline，并把专项状态从“已收口”拉回到继续 burn-down 的真实阶段。
+- [x] rust-structural-slicing-module-actions (PRD-ENGINEERING-R1200-002/004/005) [test_tier_required]: 将 `runtime/world/module_actions` 从 `include!/impl_part` 切换为目录模块，拆出 `artifact_actions` / `release_actions` / `release_normalization` / `release_support`，把 `artifact_actions.rs` 压到 1019 行，并同步退役 4 条 structural slicing frozen baseline。 Trace: .pm/tasks/task_d8ad79d7c10646008fb684934160c0b7.yaml
 
 ## 依赖
 - `doc/engineering/prd.md`
@@ -33,11 +34,12 @@
 - Batch B/C/D (`TASK-ENGINEERING-054~057`) 依赖 `TASK-ENGINEERING-052/053` 先把门禁、基线与完成态建立起来。
 
 ## 状态
-- 更新日期: 2026-04-09
+- 更新日期: 2026-04-17
 - 当前阶段: active
-- 当前任务: `TASK-ENGINEERING-104` 已完成，恢复了 `rust-oversized-file-baseline.tsv` 与 required gate 的一致性；当前 frozen baseline 为 12 个生产文件 / 7 个测试文件，专项重新回到持续 burn-down 状态。
+- 当前任务: `rust-structural-slicing-module-actions` 已完成，`runtime/world/module_actions` 已从 `include!/impl_part` 切换为目录模块，当前 frozen structural slicing baseline 已从 49 个 `slice_file` / 48 个 `include_target` 降到 47 / 46。
 - 阻塞项: 无新的脚本阻断；当前真实约束是仓库仍存在 12 个生产 Rust 超限文件与 7 个测试 Rust 超限文件，后续治理不得再宣称 `oversized code files=0, test files=0` 已达成。
 - 最新完成:
+  - `rust-structural-slicing-module-actions`：`crates/oasis7/src/runtime/world/module_actions` 已改为目录模块，原 `module_actions_impl_part1.rs` / `module_actions_impl_part2.rs` / `include!` 入口全部退役；新增 `release_normalization.rs` 承接 release 归一化与 shadow-hash helper，`artifact_actions.rs` 降到 1019 行。`env -u RUSTC_WRAPPER cargo check -p oasis7`、`env -u RUSTC_WRAPPER cargo test -p oasis7 runtime::tests::module_action_loop:: -- --nocapture`、`./scripts/check-rust-file-size.sh` 与 `git diff --check` 通过。
   - `TASK-ENGINEERING-104`：确认 `doc/.governance/rust-oversized-file-baseline.tsv` 在 `2026-03-30` 被提交为只剩注释的空壳，导致 `scripts/check-rust-file-size.sh` 在 required gate 中把全部存量超限项误判为“相对 HEAD 新增”；现已按当前仓库实况重写 baseline，恢复 `scripts/ci-tests.sh required` 的 frozen baseline 语义，并把专项状态改回继续 burn-down。
   - `TASK-ENGINEERING-054`：`oasis7_chain_runtime.rs` 拆出 `cli.rs`，`execution_bridge.rs` 迁移为目录模块，门禁基线已退休旧入口超限项。
   - `TASK-ENGINEERING-055`：`viewer/runtime_live.rs` 与测试集拆为目录模块；同时把 `runtime/events.rs`、`state.rs`、`apply_domain_event_*`、`world/event_processing.rs` 压回 1200 行内，清零本批新增 runtime 超限。
